Rename tools/webpack-make.js to ./build.js
This is our primary and only interaction point with building bundles, so it
should be in the top-level directory. Also, we are going to switch to esbuild
soon, so naming it "webpack" will be confusing. Move it to ./build.js like
cockpit-podman already did.

Keep tools/webpack-{make.js,watch} as a shell shim for a little while, which
points to the new tool and the documentation.
