+++ pkg/Makefile.am
@@ -37,7 +37,7 @@ V_BUNDLE_0 = @echo "  BUNDLE   dist";
 # but this is just a representative for all of dist/*
 $(DIST_STAMP): $(srcdir)/package-lock.json $(PKG_INPUTS)
 	@rm -f $(DIST_STAMP)
-	$(V_BUNDLE) cd $(srcdir) && NODE_ENV='$(NODE_ENV)' tools/termschutz tools/webpack-make.js
+	$(V_BUNDLE) cd $(srcdir) && NODE_ENV='$(NODE_ENV)' tools/termschutz ./build.js
 
 EXTRA_DIST += webpack.config.js package.json package-lock.json
